psa: Build Matter CryptoPAL using PSA in Open IoT SDK
This commit allows the selection of PSA as the cryptographic
algorithm used when building Matter CryptoPAL with the Open
IoT SDK.

Other important changes:

src/crypto/crypto.gni has been updated to remove the assert
added by https://github.com/project-chip/connectedhomeip/pull/23538.
This assert assumes that you only build mbedtls externally if you
have chip_crypto = 'mbedtls'. This is an invalid assumption and
is certainly not the case for the Open IoT SDK.

A new workflow has been added to the GitHub CI which runs
on each push to a PR. It is a copy of the current workflow
which builds and tests the Open IoT SDK examples and unit
tests, but with the addition of selecting PSA as the
cryptographic algorithm.

Signed-off-by: Anna Bridge <anna.bridge@arm.com>