[Fix] Fix LwIP structure initialization in unit-tests
Zero-initialization of ip6_addr_t structure in TestInetAddress.
That protects against error that may occur when the memory is aligned
to multiples of 4 and LWIP_IPV6_SCOPES is enabled. In this case,
 the ip6_addr_t contains 16 bytes array of address and 1 byte of zone
but the structure length is 20. If we don't set the initial value
for ip_addr_1 and ip_addr_2  the memcmp returns mismatch.

Signed-off-by: ATmobica <artur.tynecki@arm.com>