Tools: Comment out default Fastmodels mount link
There is a default fast models mount link in the devcontainer which
doesn't work out of the box and is there primarily as an example
of how to set it up. This can cause the devcontainer to fail to load
properly without errors. This commit comments out the link.

The commit also adds a check to the examples script to check that
the GDB plugin is actually configured.

Signed-off-by: Anna Bridge <anna.bridge@arm.com>

fix

Signed-off-by: Anna Bridge <anna.bridge@arm.com>