[OIS] Create common cmake source in platform config
Add Cmake files that create GN args file and build
the CHIP library with ExternalProject.
At the end the 'chip' is defined. It exposes CHIP headers
and libraries to the application.
Add CMake utilities for managing and retrieving build configuration.
Add Python script to generate the args.gn file in a correct format.

Signed-off-by: ATmobica <artur.tynecki@arm.com>