crypto: Keep PSA private keys in secure storage
To keep keys secure, when PSA is used. It is recommended to keep the
private key in the secure partition; in the secure storage. Keys are then
referred to with an ID rather than a value.

This requires few changes regards to how keys are copied and imported.
Previously the serialize/deserialize methods were used for that effect.

This patch introduces two specific operations for a P256Keypair:
- Copy: copy a keypair into another keypair.
- ImportRawKeypair: Import a keypair from raw bytes. Such function
was already defined in some of the unit tests as LoadKeypairFromRaw. It
is now rationalised and available from the P256KeyPair class.

The functions serialize/deserialize when implemented with PSA do not
get the raw keypair. They rather serialize the reference ID of the
private key and the public key.

The internal mechanism to commit operational fabric key has been
reworked in the PSA implementation to fix an issue where the pending
keypair is reverted.

Signed-off-by: Vincent Coubard <vincent.coubard@arm.com>